# openapi/ema.v1.s16a.json
{
...
  "info": {
    "title": "Ema local API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs/{job_id}/invoices": {
      "get": {
        "operationId": "handler_jobs__job_id__invoices_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "batch_id": "batch-exemplu",
                  "identity": {
                    "batch_id": "batch-exemplu",
                    "candidate": {
                      "client_id": "client-exemplu",
                      "cui": "00000000",
                      "pod": "POD-EXEMPLU"
                    },
                    "confirmed": false,
                    "evidence_ids": [
                      "evidence-exemplu"
                    ]
                  },
                  "missing_months": [
                    "2026-02"
                  ],
                  "rows": [
                    {
                      "anomalies": [],
                      "consumption_kwh": 1200,
                      "id": "invoice-exemplu",
                      "month": "2026-01",
                      "source_evidence_ids": [
                        "evidence-exemplu"
                      ]
                    }
                  ]
                },
                "schema": {
                  "properties": {
                    "batch_id": {
                      "type": "string"
                    },
                    "identity": {
                      "properties": {
                        "batch_id": {
                          "type": "string"
                        },
                        "candidate": {
                          "properties": {
                            "client_id": {
                              "type": "string"
                            },
                            "cui": {
                              "type": "string"
                            },
                            "pod": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "client_id",
                            "cui",
                            "pod"
                          ],
                          "type": "object"
                        },
                        "confirmed": {
                          "type": "boolean"
                        },
                        "evidence_ids": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        }
                      },
                      "required": [
                        "batch_id",
                        "candidate",
                        "confirmed",
                        "evidence_ids"
                      ],
                      "type": "object"
                    },
                    "missing_months": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "rows": {
                      "items": {
                        "properties": {
                          "anomalies": {
                            "items": {},
                            "type": "array"
                          },
                          "consumption_kwh": {
                            "type": "integer"
                          },
                          "id": {
                            "type": "string"
                          },
                          "month": {
                            "type": "string"
                          },
                          "source_evidence_ids": {
                            "items": {
                              "type": "string"
                            },
                            "type": "array"
                          }
                        },
                        "required": [
                          "id",
                          "month",
                          "consumption_kwh",
                          "source_evidence_ids",
                          "anomalies"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "batch_id",
                    "identity",
                    "rows",
                    "missing_months"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      }
    },
    "/jobs/{job_id}/invoices/identity": {
      "get": {
        "operationId": "handler_jobs__job_id__invoices_identity_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "batch_id": "batch-exemplu",
                  "candidate": {
                    "client_id": "client-exemplu",
                    "cui": "00000000",
                    "pod": "POD-EXEMPLU"
                  },
                  "confirmed": false,
                  "evidence_ids": [
                    "evidence-exemplu"
                  ]
                },
                "schema": {
                  "properties": {
                    "batch_id": {
                      "type": "string"
                    },
                    "candidate": {
                      "properties": {
                        "client_id": {
                          "type": "string"
                        },
                        "cui": {
                          "type": "string"
                        },
                        "pod": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "client_id",
                        "cui",
                        "pod"
                      ],
                      "type": "object"
                    },
                    "confirmed": {
                      "type": "boolean"
                    },
                    "evidence_ids": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "batch_id",
                    "candidate",
                    "confirmed",
                    "evidence_ids"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      },
      "post": {
        "operationId": "handler_jobs__job_id__invoices_identity_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "client_id": "client-exemplu",
                "confirm": true,
                "on_revision": 1
              },
              "schema": {
                "properties": {
                  "client_id": {
                    "type": "string"
                  },
                  "confirm": {
                    "type": "boolean"
                  },
                  "on_revision": {
                    "type": "integer"
                  }
                },
                "required": [
                  "client_id",
                  "on_revision",
                  "confirm"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "batch_id": "batch-exemplu",
                  "candidate": {
                    "client_id": "client-exemplu",
                    "cui": "00000000",
                    "pod": "POD-EXEMPLU"
                  },
                  "confirmed": true,
                  "evidence_ids": [
                    "evidence-exemplu"
                  ]
                },
                "schema": {
                  "properties": {
                    "batch_id": {
                      "type": "string"
                    },
                    "candidate": {
                      "properties": {
                        "client_id": {
                          "type": "string"
                        },
                        "cui": {
                          "type": "string"
                        },
                        "pod": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "client_id",
                        "cui",
                        "pod"
                      ],
                      "type": "object"
                    },
                    "confirmed": {
                      "type": "boolean"
                    },
                    "evidence_ids": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "batch_id",
                    "candidate",
                    "confirmed",
                    "evidence_ids"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "501": {
            "description": "Contract only; use case pending"
          }
        },
        "summary": "Handler",
        "tags": [
          "provisional"
        ],
        "x-provisional": true
      }
    },
...
  }
}